:sparkles: ストーリー画面を作成した (#37)
