+++ lib/constants/assets.dart
@@ -1,12 +1,11 @@
 class AppAssets {
-
   /// アイコンへのベースパス
   static const String _background = 'assets/images/background';
   static const String _character = 'assets/images/character';
   static const String _icon = 'assets/images/icon';
 
   // --- Background ---
-  static const String backgroundClassroom = '$_background/classroom.png';
+  static const String backgroundClassroom = '$_background/classroom.jpg';
   // --- Character ---
   static const String characterSuzunari = '$_character/suzunari.png';
   static const String characterComingsoon = '$_character/comingsoon.webp';
@@ -14,4 +13,10 @@ class AppAssets {
   static const String iconKiroku = '$_icon/kiroku.png';
   static const String iconLockClosed = '$_icon/lock_closed.png';
   static const String iconLockOpen = '$_icon/lock_open.png';
+
+  /// BGMへのベースパス
+  static const String _bgm = 'assets/bgm';
+
+  // BGM集
+  static const String sadBGM = '$_bgm/sad.mp3';
 }
