+++ lib/stories/suzunari_oto.dart
@@ -1,0 +1,31 @@
+// 1行22文字以内
+// 3行まで
+class EpisodeSuzunariOto {
+  static const List<String> suzunariOtoStory = [
+    'おと「ここ、座ってもいいですか？」',
+    '図書館のすみにある静かな席\nノートに目を落としていた俺は 声に気づいて顔を上げた',
+    'そこにいたのは 見たことのない女の子\nたぶん1年生 髪をゆるく結んで セーラー服に小さな鈴のキーホルダーが揺れてた',
+    '「…いいけど」',
+    '「ありがとうございます」',
+    'にこっと笑って 彼女は俺の隣に座った\nそして静かにノートを広げて 鉛筆を走らせる',
+    '「ここ、静かでいいですね」',
+    '「まあ…いつも人いないから」',
+    '「落ち着く場所って 大事ですよね」',
+    'なんて答えればいいのかわからず 俺は「うん」とだけ返した\nそれでも彼女はにこにこしていた',
+    '「先輩って 2年生ですか？」',
+    '「うん」',
+    '「ですよね わたし1年です 鈴鳴おとって言います」',
+    '「俺は○○」',
+    '「〇〇せんぱい」',
+    '呼ばれて 少し胸が高まった\nけど 不思議と嫌じゃなかった',
+    '「ここ、毎日来てもいいですか？」',
+    '「…好きにすれば」',
+    '「じゃあ 来ます」',
+    'それだけ言って 鈴鳴はまたノートに集中しはじめた\n隣からシャーペンのカリカリという音だけが聞こえる',
+    'その日から 時々 彼女は来るようになった\nおしゃべりは多くないけど 目が合うと必ずにこって笑う',
+    'それだけで 何も言わなくても時間が流れた',
+    '梅雨に入ったころ 鈴鳴はふいに来なくなった',
+    '理由もわからないまま 俺はまた一人で 窓の外を眺めるようになった',
+    'けれど 時々風の音に混じって 鈴の音が聞こえた気がした',
+  ];
+}
