+++ pubspec.yaml
@@ -69,6 +69,7 @@ flutter:
     - assets/images/background/
     - assets/images/character/
     - assets/images/icon/
+    - assets/bgm/
 
   # To add assets to your application, add an assets section, like this:
   # assets:
